+++ CHANGELOG.md
@@ -10,6 +10,7 @@ IMPROVEMENTS:
 * agent: Reduced the timeouts for the `-dev` server mode so that the development server starts up almost instantly. [GH-2984]
 * agent: Added `verify_incoming_rpc` and `verify_incoming_https` options for more granular control over incoming TLS enforcement. [GH-2974]
 * agent: Updated parts of Consul's Docker library in order to fix the build on Illumos. [GH-2989]
+* agent: Use bind address as source for outgoing connections. [GH-2822]
 * api: Added the ACL replication status endpoint to the Go API client library. [GH-2947]
 * cli: Added Raft protocol version to output of `operator raft list-peers` command.[GH-2929]
 * ui: Added optional JSON validation when editing KV entries in the web UI. [GH-2712]
@@ -17,6 +18,7 @@ IMPROVEMENTS:
 BUG FIXES:
 
 * server: Fixed a panic when the tombstone garbage collector was stopped. [GH-2087]
+* server: Fixed a panic in Autopilot that could occur when a node is elected but cannot complete leader establishment and steps back down. [GH-2980]
 * ui: Add and update favicons [GH-2945]
 
 ## 0.8.1 (April 17, 2017)